+++ jest.config.js
@@ -18,10 +18,10 @@ module.exports = {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 87,
-      functions: 100,
-      lines: 92,
-      statements: 92,
+      branches: 69,
+      functions: 79,
+      lines: 73,
+      statements: 73,
     },
   },
   // A preset that is used as a base for Jest's configuration