+++ src/feature-toggle-provider.tsx
@@ -1,14 +1,42 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import React, { ReactNode, FC, ReactElement } from 'react';
-import { set, FeatureToggleServiceConfig } from 'feature-toggle-service';
+import React, { createContext, useContext, ReactNode } from 'react';
+import { set, isOn, FeatureToggleServiceConfig } from 'feature-toggle-service';
 
 export interface ReactorFeatureToggleProviderProps {
   children: ReactNode;
   featureToggleService: FeatureToggleServiceConfig;
 }
+interface FeatureToggleContextType {
+  isOn: (featureName: string) => boolean;
+  toggles: FeatureToggleServiceConfig | null;
+}
 
-export const FeatureToggleProvider: FC<ReactorFeatureToggleProviderProps> = props => {
-  set(props.featureToggleService);
+const FeatureToggleContext = createContext<FeatureToggleContextType>({
+  isOn,
+  toggles: null,
+});
 
-  return props.children as ReactElement<any>;
+export const FeatureToggleProvider = ({
+  children,
+  featureToggleService,
+}: ReactorFeatureToggleProviderProps) => {
+  set(featureToggleService);
+
+  return (
+    <FeatureToggleContext.Provider
+      value={{ isOn, toggles: featureToggleService }}
+    >
+      {children}
+    </FeatureToggleContext.Provider>
+  );
 };
+
+export function useFeatureToggle() {
+  const context = useContext(FeatureToggleContext);
+  if (!context) {
+    throw new Error(
+      'useFeatureToggle must be used within FeatureToggleClientProvider'
+    );
+  }
+
+  return context;
+}