+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Added
+
+- Adding `useFeatureToggle()` context
+
 ## [6.0.1][] - 2025-12-04
 
 ### Fixed